// src/app/api-swagger/models/collection-point-dto.ts
/* tslint:disable */
/* eslint-disable */
export interface CollectionPointDto {
  consumptionHighTarif?: null | number;
  consumptionLowTarif?: null | number;
  costAfter?: null | number;
  costBefor?: null | number;
  costHighTarif?: null | number;
  costLowTarif?: null | number;
  distributionFactor?: null | number;
  economicModelCalculationId?: null | number;
  fixedMonthlyPayment?: null | number;
  id?: number;
  name?: null | string;
  saving?: null | number;
}