// src/app/api-swagger/models/economic-model-calculation-create-req.ts
/* tslint:disable */
/* eslint-disable */
import { CollectionPointDto } from './collection-point-dto';
import { PvType } from './pv-type';
export interface EconomicModelCalculationCreateReq {
  azimut?: null | number;
  collectionPoints?: null | Array<CollectionPointDto>;
  communityName?: null | string;
  grant?: null | number;
  lifetime?: null | number;
  locationGps?: null | string;
  pvPeakPower?: null | number;
  pvSystemCost?: null | number;
  pvType?: PvType;
  slope?: null | number;
  userId?: null | number;
}